Code cleanup
Set lang version to 10.
Set global usings
File scoped namespaces
Pattern match null checks
Fixed some comments to use XML comments
