+++ IntelliTect.TestTools.Console.Tests/ConsoleAssertTests.cs
@@ -1,188 +1,185 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
-using System;
-using System.Runtime.InteropServices;
+﻿using System.Runtime.InteropServices;
 
-namespace IntelliTect.TestTools.Console.Tests
+namespace IntelliTect.TestTools.Console.Tests;
+
+[TestClass]
+public class ConsoleAssertTests
 {
-    [TestClass]
-    public class ConsoleAssertTests
+    [TestMethod]
+    public void ConsoleTester_Sample_InigoMontoya()
     {
-        [TestMethod]
-        public void ConsoleTester_Sample_InigoMontoya()
-        {
-            const string view =
+        const string view =
 @"First name: <<Inigo
 >>Last name: <<Montoya
 >>Hello, Inigo Montoya.";
 
-            ConsoleAssert.Expect(view,
-            () =>
-            {
-                System.Console.Write("First name: ");
-                string fname = System.Console.ReadLine();
-
-                System.Console.Write("Last name: ");
-                string lname = System.Console.ReadLine();
-
-                System.Console.Write("Hello, {0} {1}.", fname, lname);
-            });
-        }
-
-        [TestMethod]
-        public void ConsoleTester_HelloWorld_NoInput()
+        ConsoleAssert.Expect(view,
+        () =>
         {
-            const string view = "Hello World";
+            System.Console.Write("First name: ");
+            string fname = System.Console.ReadLine();
 
-            ConsoleAssert.Expect(view, () =>
-            {
-                System.Console.Write("Hello World");
-            }, NormalizeOptions.None);
-        }
+            System.Console.Write("Last name: ");
+            string lname = System.Console.ReadLine();
 
-        [TestMethod]
-        public void ConsoleTester_HelloWorld_TrimLF()
+            System.Console.Write("Hello, {0} {1}.", fname, lname);
+        });
+    }
+
+    [TestMethod]
+    public void ConsoleTester_HelloWorld_NoInput()
+    {
+        const string view = "Hello World";
+
+        ConsoleAssert.Expect(view, () =>
         {
-            const string view = "Hello World\n";
+            System.Console.Write("Hello World");
+        }, NormalizeOptions.None);
+    }
 
-            ConsoleAssert.Expect(view, () =>
-            {
-                System.Console.WriteLine("Hello World");
-            });
-        }
+    [TestMethod]
+    public void ConsoleTester_HelloWorld_TrimLF()
+    {
+        const string view = "Hello World\n";
 
-        [TestMethod]
-        [DataRow("\u001b[49mMontoya", "Montoya", true)]
-        [DataRow("Inigo\u001b[49mMontoya", "InigoMontoya", true)]
-        [DataRow("Inigo\u001b[49m", "Inigo", true)]
-        [DataRow("\u001b[49m", "", true)]
-        [DataRow("\u001b[101mMontoya", "Montoya", true)]
-        [DataRow("\u001b[101mMontoya", "\u001b[101mMontoya", false)]
-        public void ConsoleTester_StringWithVT100Characters_VT100Stripped(string input,
-            string expected,
-            bool stripVT100)
+        ConsoleAssert.Expect(view, () =>
         {
-            ConsoleAssert.Expect(expected, () =>
-            {
-                System.Console.WriteLine(input);
-            });
-        }
+            System.Console.WriteLine("Hello World");
+        });
+    }
 
-        [TestMethod]
-        public void ConsoleTester_ExplicitStrippingExplicitly_VT100Stripped()
+    [TestMethod]
+    [DataRow("\u001b[49mMontoya", "Montoya", true)]
+    [DataRow("Inigo\u001b[49mMontoya", "InigoMontoya", true)]
+    [DataRow("Inigo\u001b[49m", "Inigo", true)]
+    [DataRow("\u001b[49m", "", true)]
+    [DataRow("\u001b[101mMontoya", "Montoya", true)]
+    [DataRow("\u001b[101mMontoya", "\u001b[101mMontoya", false)]
+    public void ConsoleTester_StringWithVT100Characters_VT100Stripped(string input,
+        string expected,
+        bool stripVT100)
+    {
+        ConsoleAssert.Expect(expected, () =>
         {
-            string input = "\u001b[49mMontoya";
-            string expected = "Montoya";
+            System.Console.WriteLine(input);
+        });
+    }
 
-            ConsoleAssert.Expect(expected, () =>
-            {
-                System.Console.Write(input);
-            }, NormalizeOptions.StripAnsiEscapeCodes);
-        }
+    [TestMethod]
+    public void ConsoleTester_ExplicitStrippingExplicitly_VT100Stripped()
+    {
+        string input = "\u001b[49mMontoya";
+        string expected = "Montoya";
 
-        [TestMethod]
-        public void GivenStringLiteral_ExpectedOutputNormalized_OutputMatches()
+        ConsoleAssert.Expect(expected, () =>
         {
-            const string view = @"Begin
+            System.Console.Write(input);
+        }, NormalizeOptions.StripAnsiEscapeCodes);
+    }
+
+    [TestMethod]
+    public void GivenStringLiteral_ExpectedOutputNormalized_OutputMatches()
+    {
+        const string view = @"Begin
 Middle
 End";
+        ConsoleAssert.Expect(view, () =>
+        {
+            System.Console.WriteLine("Begin");
+            System.Console.WriteLine("Middle");
+            System.Console.WriteLine("End");
+        });
+    }
+
+    [TestMethod]
+    public void ConsoleTester_HelloWorld_TrimCRLF()
+    {
+        const string view = "Hello World";
+
+        ConsoleAssert.Expect(view, () =>
+        {
+            System.Console.Write("Hello World");
+        });
+    }
+
+    [TestMethod]
+    public void ConsoleTester_HelloWorld_DontNormalizeCRLF()
+    {
+        const string view = "Hello World\r\n";
+
+        Assert.ThrowsException<Exception>(() =>
+        {
             ConsoleAssert.Expect(view, () =>
             {
-                System.Console.WriteLine("Begin");
-                System.Console.WriteLine("Middle");
-                System.Console.WriteLine("End");
-            });
-        }
+                System.Console.Write("Hello World\r");
+            }, NormalizeOptions.None);
+        });
+    }
 
-        [TestMethod]
-        public void ConsoleTester_HelloWorld_TrimCRLF()
+    [TestMethod]
+    [DataRow("C++")]
+    [DataRow("word + word")]
+    [DataRow("+hello+world+")]
+    public void ConsoleTester_OutputIncludesPluses_PlusesAreNotStripped(string consoleInput)
+    {
+        Exception exception = Assert.ThrowsException<Exception>(() =>
         {
-            const string view = "Hello World";
-
-            ConsoleAssert.Expect(view, () =>
+            ConsoleAssert.Expect(consoleInput, () =>
             {
-                System.Console.Write("Hello World");
-            });
-        }
+                System.Console.Write(""); // Always fail
+            }, NormalizeOptions.None);
+        });
+        StringAssert.Contains(exception.Message, consoleInput);
+    }
 
-        [TestMethod]
-        public void ConsoleTester_HelloWorld_DontNormalizeCRLF()
-        {
-            const string view = "Hello World\r\n";
-
-            Assert.ThrowsException<Exception>(() =>
-            {
-                ConsoleAssert.Expect(view, () =>
-                {
-                    System.Console.Write("Hello World\r");
-                }, NormalizeOptions.None);
-            });
-        }
-
-        [TestMethod]
-        [DataRow("C++")]
-        [DataRow("word + word")]
-        [DataRow("+hello+world+")]
-        public void ConsoleTester_OutputIncludesPluses_PlusesAreNotStripped(string consoleInput)
-        {
-            Exception exception = Assert.ThrowsException<Exception>(() =>
-            {
-                ConsoleAssert.Expect(consoleInput, () =>
-                {
-                    System.Console.Write(""); // Always fail
-                }, NormalizeOptions.None);
-            });
-            StringAssert.Contains(exception.Message, consoleInput);
-        }
-
-        [TestMethod]
-        public void ConsoleTester_HelloWorld_MissingNewline()
-        {
-            const string view = @"Hello World
+    [TestMethod]
+    public void ConsoleTester_HelloWorld_MissingNewline()
+    {
+        const string view = @"Hello World
 ";
 
-            ConsoleAssert.Expect(view, () =>
-            {
-                System.Console.WriteLine("Hello World");
-            }, NormalizeOptions.None);
-        }
-
-        [TestMethod]
-        public void ExecuteProcess_PingLocalhost_Success()
+        ConsoleAssert.Expect(view, () =>
         {
-            string expected = $@"*
+            System.Console.WriteLine("Hello World");
+        }, NormalizeOptions.None);
+    }
+
+    [TestMethod]
+    public void ExecuteProcess_PingLocalhost_Success()
+    {
+        string expected = $@"*
 Pinging * ?::1? with 32 bytes of data:
 Reply from ::1: time*";
-            string pingArgs = "-c 4 localhost";
-            if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-            {
-                pingArgs = pingArgs.Replace("-c ", "-n ");
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-            {
-                expected = $@"PING *(* (::1)) 56 data bytes
-64 bytes from * (::1): icmp_seq=1 ttl=64 time=* ms*";
-            }
-            else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-            {
-                expected = $@"PING *(*): 56 data bytes
-64 bytes from *: icmp_seq=? ttl=64 time=* ms*";
-            }
-
-            ConsoleAssert.ExecuteProcess(
-                expected,
-                "ping", pingArgs, out string standardOutput, out _);
-        }
-
-        [TestMethod]
-        public void ExecuteLike_GivenVariableCRLFWithNLComparedToCRNL_Success()
+        string pingArgs = "-c 4 localhost";
+        if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
         {
-            const string expected = "(abstract, 1)\n(abstract, 2)\n\n";
-            const string output = "(abstract, 1)\r\n(abstract, 2)\r\n";
-
-            IntelliTect.TestTools.Console.ConsoleAssert.ExpectLike(expected, () =>
-            {
-                System.Console.WriteLine(output);
-            });
+            pingArgs = pingArgs.Replace("-c ", "-n ");
         }
+        else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
+        {
+            expected = $@"PING *(* (::1)) 56 data bytes
+64 bytes from * (::1): icmp_seq=1 ttl=64 time=* ms*";
+        }
+        else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
+        {
+            expected = $@"PING *(*): 56 data bytes
+64 bytes from *: icmp_seq=? ttl=64 time=* ms*";
+        }
+
+        ConsoleAssert.ExecuteProcess(
+            expected,
+            "ping", pingArgs, out string standardOutput, out _);
+    }
+
+    [TestMethod]
+    public void ExecuteLike_GivenVariableCRLFWithNLComparedToCRNL_Success()
+    {
+        const string expected = "(abstract, 1)\n(abstract, 2)\n\n";
+        const string output = "(abstract, 1)\r\n(abstract, 2)\r\n";
+
+        IntelliTect.TestTools.Console.ConsoleAssert.ExpectLike(expected, () =>
+        {
+            System.Console.WriteLine(output);
+        });
     }
 }
