+++ IntelliTect.TestTools.Console.sln
@@ -1,19 +1,21 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 16
-VisualStudioVersion = 16.0.30114.105
+# Visual Studio Version 17
+VisualStudioVersion = 17.3.32825.248
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelliTect.TestTools.Console.Tests", "IntelliTect.TestTools.Console.Tests\IntelliTect.TestTools.Console.Tests.csproj", "{EA67850E-E84F-4006-85F8-5D91FBCA33E1}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntelliTect.TestTools.Console.Tests", "IntelliTect.TestTools.Console.Tests\IntelliTect.TestTools.Console.Tests.csproj", "{EA67850E-E84F-4006-85F8-5D91FBCA33E1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "IntelliTect.TestTools.Console", "IntelliTect.TestTools.Console\IntelliTect.TestTools.Console.csproj", "{16FF13CC-0EE2-4D04-932D-BCD939F5DEF4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntelliTect.TestTools.Console", "IntelliTect.TestTools.Console\IntelliTect.TestTools.Console.csproj", "{16FF13CC-0EE2-4D04-932D-BCD939F5DEF4}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{FEFBB94E-0375-412A-8948-ED2EEFD99A0D}"
+	ProjectSection(SolutionItems) = preProject
+		Directory.Build.props = Directory.Build.props
+	EndProjectSection
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{EA67850E-E84F-4006-85F8-5D91FBCA33E1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -24,5 +26,11 @@ Global
 		{16FF13CC-0EE2-4D04-932D-BCD939F5DEF4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{16FF13CC-0EE2-4D04-932D-BCD939F5DEF4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{16FF13CC-0EE2-4D04-932D-BCD939F5DEF4}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {48D966FF-61E6-4B87-AF6F-0BBDA936D576}
 	EndGlobalSection
 EndGlobal
